+++ unittests/test_virtual_machine.py
@@ -1,5 +1,5 @@
 import unittest
-from spinn_machine.virutal_machine import VirtualMachine
+from spinn_machine.virtual_machine import VirtualMachine
 
 
 class TestVirtualMachine(unittest.TestCase):
